fix: 로그인 로직 수정
루트 / 경로로 오면 토큰 확인 후 토큰이 있으면 /main 으로 이동
토큰이 없으면 /login 으로 이동
/login 에서는 구글 로그인 여부 확인하고 구글 로그인 성공했으면 로컬 서버에 로그인 시도
로그인 성공하면 jwt 받아서 로컬 스토리지에 등록
로그인 실패하면 회원가입 후 jwt 받아서 로컬스토리지에 등록
유저 정보 context api에 등록
